refactor: move registry logic to separate module
